+++ pkg/system/gateway/internal/test/ac/cmd/cmd.go
@@ -1,10 +1,10 @@
 package main
 
 import (
-	"github.com/vanti-dev/sc-bos/pkg/system/proxy/internal/test/hub"
+	"github.com/vanti-dev/sc-bos/pkg/system/gateway/internal/test/ac"
 )
 
 func main() {
 	// Important: don't add code to this file, or you'll risk breaking test result caching.
-	hub.Main()
+	ac.Main()
 }
