+++ pkg/system/gateway/internal/test/testdata/ac1/app.conf.json
