+++ pkg/system/gateway/internal/test/testdata/ac1/system.conf.json
