+++ pkg/system/gateway/internal/test/testdata/ac2/app.conf.json
