+++ pkg/system/gateway/internal/test/testdata/ac2/system.conf.json
