+++ pkg/system/gateway/internal/test/testdata/gw1/app.conf.json
