+++ pkg/system/gateway/internal/test/testdata/gw1/system.conf.json
@@ -2,7 +2,7 @@
   "name": "gw1",
   "logger": {"initialFields": {"proc": "gw1"}},
   "systems": {
-    "proxy": {},
+    "gateway": {},
     "hub": {"storage": {"type": "proxy"}},
     "tenants": {"storage": {"type": "proxy"}}
   }
