+++ pkg/system/gateway/internal/test/testdata/gw2/app.conf.json
