+++ pkg/system/gateway/internal/test/testdata/gw2/system.conf.json
@@ -2,7 +2,7 @@
   "name": "gw2",
   "logger": {"initialFields": {"proc": "gw2"}},
   "systems": {
-    "proxy": {},
+    "gateway": {},
     "hub": {"storage": {"type": "proxy"}},
     "tenants": {"storage": {"type": "proxy"}}
   }
