+++ pkg/system/gateway/internal/test/testdata/hub/app.conf.json
