+++ pkg/system/gateway/internal/test/testdata/hub/system.conf.json
